drobemate_backend/requirements.txt
asgiref==3.11.0
Django==4.2.26
djangorestframework==3.16.1
dotenv==0.9.9
pillow==11.3.0
psycopg==3.2.12
python-dotenv==1.2.1
sqlparse==0.5.3
typing_extensions==4.15.0
